Markdown converter to TS
